Move testfiles to top level
